+++ checkout/Changelog.md
@@ -4,3 +4,6 @@
 
 # 18. December 2019
 * Reduce calls for updating items in `SetSourcesForCartItems`
+
+# 7. January 2020
+* Generate a new Idempotency Key in the PaymentSelection if an payment error occurs (canceled / aborted by customer) to allow the customer to retry